requirements.txt
Flask==0.12.2
numpy==1.14.0
pyramid-arima==0.6.2
scipy==0.19.1
pandas==0.20.3
rpy2==2.9.2